+++ JuliaLibWrapping/test/expected_carray_owned_lowlevel.py
@@ -56,7 +56,7 @@ _jlw_loaded.add(_jlw_this_pkg)
 
 class CString_owned(ctypes.Structure):
     _fields_ = [
-        ("length", ctypes.c_int32),
+        ("length", ctypes.c_int64),
         ("data", ctypes.POINTER(ctypes.c_uint8)),
     ]
 
@@ -93,7 +93,7 @@ class CString_owned(ctypes.Structure):
 
 class CVector_owned_Float64(ctypes.Structure):
     _fields_ = [
-        ("dims", (ctypes.c_int32 * 1)),
+        ("dims", (ctypes.c_int64 * 1)),
         ("data", ctypes.POINTER(ctypes.c_double)),
     ]
